Require (again) a Horde_Mime_Part object when creating
Bite the bullet and required the Horde_Mime_Part object when creating a
Horde_Mime_Viewer instance. Especially for embedded parts, or for
multiparts that may have to recursively look through the subparts, the
information stored in the MIME part may be needed for things like
determining whether a driver can display a part with the given mode.