Bug: 14324  Be more graceful with rfc822/message attachment failures.
If the message fails to send, check that it's not because of a broken
rfc822/message attachement - most likely a too long header line. This is
done by re-encoding the attachment as a base64 encoded
application/octet-stream part. Not ideal, but we don't want to alter the
original message in anyway, and this at least allows it to be forwarded.